Refactor transformation functions into a modular package structure with backward compatibility
